+++ packages/gallery/src/settings/options/galleryMargin.js
@@ -3,7 +3,7 @@ import { INPUT_TYPES } from '../utils/constants';
 export default {
   title: 'Gallery Spacing',
   description: `Set the margin between the gallery and it's container.`,
-  isRelevant: (styleParams) => true,
+  isRelevant: () => true,
   type: INPUT_TYPES.NUMBER,
   default: 0,
 }