+++ packages/gallery/src/settings/options/slideshowInfoSize.js
@@ -3,7 +3,7 @@ import { INPUT_TYPES } from '../utils/constants';
 export default {
   title: 'Slideshow Info Section Size',
   description: `Set the size of the info section in slideshow. the item's height will be adjusted accordingly`,
-  isRelevant: (styleParams) => true,
+  isRelevant: () => true,
   type: INPUT_TYPES.NUMBER,
   default: 200,
 }