+++ packages/gallery/src/settings/options/videoPlay.js
@@ -4,7 +4,7 @@ import { createOptions } from '../utils/utils';
 export default {
   title: 'Playing Options',
   description: `Choose when video items will start playing: On hover, on click or autoplay`,
-  isRelevant: (styleParams) => true,
+  isRelevant: () => true,
   options: createOptions('videoPlay'),
   type: INPUT_TYPES.OPTIONS,
   default: GALLERY_CONSTS.videoPlay.HOVER,