+++ packages/gallery/tests/e2e/styleParams/imageMargin.e2e.spec.js
@@ -1,8 +1,6 @@
 import GalleryDriver from '../../drivers/pptrDriver';
 import {toMatchImageSnapshot} from '../../drivers/matchers';
 
-jest.setTimeout(30000)
-
 expect.extend({ toMatchImageSnapshot });
 
 describe('imageMargin - e2e', () => {