+++ packages/gallery/tests/e2e/styleParams/smartCrop.e2e.spec.js
@@ -2,8 +2,6 @@ import GalleryDriver from '../../drivers/pptrDriver';
 import GALLERY_CONSTS from '../../../src/common/constants';
 import {toMatchImageSnapshot} from '../../drivers/matchers';
 
-jest.setTimeout(30000)
-
 expect.extend({ toMatchImageSnapshot });
 
 describe('smartCrop - e2e', () => {